Lord help this computer to not die

+++ CosMtx.jl
@@ -7,11 +7,11 @@ function CosMtx(D,A,n,F)
     
     hc=BigFloat(1)/BigFloat((n-1));
     Coarse=collect(BigFloat,-D+hc:hc:D)
-    Coarse=Coarse.-(5/90)
+    Coarse=Coarse.-(BigFloat(5)/BigFloat(90))
     LC=length(Coarse)
     hf=BigFloat(1)/(BigFloat((n-1))*BigFloat(F));
     Fine=collect(BigFloat,-D+hf:hf:D)
-    Fine=Fine.-(5/900);
+    Fine=Fine.-(BigFloat(5)/BigFloat(900));
     F1=collect(BigFloat,A:hf:D-A);
     SB=length(F1);
     LF=length(Fine)
